nginx: don't include the OGO header in the API rewrite calls
OGO has a mechanism to prevent request looping on their server: if a
request comes back with a certain header stamped (X-OGO-Shield), they
abort the request with a code 508 since it's come up before.

In our case, we're fine because it's not exactly the same request:

1. the backend receives a call on `mobilic.beta.gouv.fr/api`
2. re-issues the same request on `api.mobilic.beta.gouv.fr`.

The engineer at OGO (thanks Arnaud) suggested removing this custom
header to make sure OGO doesn't break the request chain with a 508
error.